+++ atecAPP-server/src/main/java/com/atec/exception/ClientNotFoundException.java
@@ -1,4 +1,4 @@
-package com.atec.service;
+package com.atec.exception;
 
 public class ClientNotFoundException extends RuntimeException {
     public ClientNotFoundException(String s){
