+++ atecAPP-server/src/main/java/com/atec/exception/EquipmentNotFoundException.java
@@ -1,4 +1,4 @@
-package com.atec.service;
+package com.atec.exception;
 
 public class EquipmentNotFoundException extends RuntimeException {
     public EquipmentNotFoundException(String s){
