+++ atecAPP-server/src/main/java/com/atec/service/ClientService.java
@@ -3,8 +3,8 @@ package com.atec.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.atec.exception.ClientNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
