+++ atecAPP-server/src/main/java/com/atec/service/EquipmentService.java
@@ -1,13 +1,13 @@
 package com.atec.service;
 
 import com.atec.dto.EquipmentDTO;
+import com.atec.exception.EquipmentNotFoundException;
 import com.atec.mapper.EquipmentMapper;
 import com.atec.model.Client;
 import com.atec.model.Equipment;
 import com.atec.repository.ClientRepository;
 import com.atec.repository.EquipmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
